fix(rp2350): use software spinlocks due to erratum RP2350-E2
The runtime spinlocks (printLock, schedulerLock, atomicsLock, futexLock)
used the hardware SIO spinlocks on both RP2040 and RP2350. Erratum
RP2350-E2 makes the SIO spinlocks unreliable on the RP2350: a core can
observe a lock as acquired when it is not, which corrupts scheduler
state when running with -scheduler=cores. In practice core 1 jumps
through a torn task pointer into data RAM and dies with an INVPC
HardFault or parks in LOCKUP shortly after startSecondaryCores.

Move the hardware spinlock implementation into runtime_rp2040.go and
give the RP2350 a software spinlock built on 32-bit atomics, matching
what the pico-sdk did for this erratum. The Cortex-M33 lowers 8/16/32
bit atomics to native ldrex/strex (no __atomic_* libcalls, verified in
the disassembly), so the lock does not recurse into atomicsLock. Lock
waits with wfe like before, and Unlock keeps the sev so cores blocked
in Lock or schedulerUnlockAndWait wake up exactly as they did with the
hardware version.

Verified on an Adafruit Metro RP2350: with the old code a cores
scheduler build never survives boot; with this change it boots, runs
goroutines on both cores in parallel, and sustains a two-core
atomics/scheduler stress test.
